fix(vue): support dots in component names
This changes the `tag_name` parsing to allow a dot in the name. So, this
will be fixed for both start tag and end tags.

Added a test that previously failed for this case:

```vue
<template>
    <Hello.World> ... </Hello.World>
</template>
```

Which was previously parsed as

```
(document
  (template_element
    (start_tag
      (tag_name))
    (element
      (start_tag
        (tag_name)
        (directive_attribute
          (directive_value)))
      (text)
      (end_tag
        (tag_name)
        (ERROR
          (UNEXPECTED 'W'))))
    (end_tag
      (tag_name))))
```

This is the same problem as described in #3 but I don't think the
existing PR fixes the problem correctly.

After this change, the whole "Hello.World" part is parsed as `tag_name`.

Running `npm run build` made some changes in `parser.c`, but I don't
know if they are important to this change. Please let me know if I
should revert those changes.

Fixes #3